Transformando os custos em inteiros e colocando os custos base.

+++ Código/GerenciadorPampatec/src/main/java/com/ideiah/gerenciadorpampatec/model/Planofinanceiro.java
@@ -17,8 +17,8 @@ public class Planofinanceiro implements java.io.Serializable {
     private String investimentoInicial;
     private Set projetos = new HashSet(0);
     private Set<Custo> custo = new HashSet<>(0);
-    private Float valorTotalFixo;
-    private Float valorTotalVariavel;
+    private Integer valorTotalFixo;
+    private Integer valorTotalVariavel;
 
     public Planofinanceiro() {
     }
@@ -92,19 +92,19 @@ public class Planofinanceiro implements java.io.Serializable {
         this.custo = custo;
     }
 
-    public Float getValorTotalFixo() {
+    public Integer getValorTotalFixo() {
         return valorTotalFixo;
     }
 
-    public void setValorTotalFixo(Float valorTotalFixo) {
+    public void setValorTotalFixo(Integer valorTotalFixo) {
         this.valorTotalFixo = valorTotalFixo;
     }
 
-    public Float getValorTotalVariavel() {
+    public Integer getValorTotalVariavel() {
         return valorTotalVariavel;
     }
 
-    public void setValorTotalVariavel(Float valorTotalVariavel) {
+    public void setValorTotalVariavel(Integer valorTotalVariavel) {
         this.valorTotalVariavel = valorTotalVariavel;
     }
